+++ CHANGELOG.md
@@ -2,9 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-04-03
+
+### Added
+
+- Explanations and examples for [Expectation Declarations], includes:
+  - Declaring method call
+  - Declaring method argument
+  - Declaring return value
+  - Declaring call count
+  - Setting public properties
+  - Multiple calls with different expectations
+  - Expectation declaration utilities
+
 ## [1.0.0] - 2025-03-22
 
 - The first release of this package.
-- Includes explanations and examples for making mock/spy objects.
+- Includes explanations and examples for [making mock/spy] objects.
 
+[1.1.0]: https://github.com/amyavari/php-mockery-examples-and-explanations/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/amyavari/php-mockery-examples-and-explanations/compare/d6e594f...v1.0.0
+[Expectation Declarations]: https://docs.mockery.io/en/stable/reference/expectations.html
+[making mock/spy]: https://docs.mockery.io/en/stable/reference/creating_test_doubles.html